pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Research agent scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "ty",
]

[project.scripts]
research-agent = "research_agent.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
